backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
scikit-learn
pandas
numpy